Introduce new helper modules for colorization, depth formatting, method name formatting, and parameter formatting; refactor BaseFormatter to extend TraceDataFormatter
