+++ lib/trace_viz/exporters/formatters/base_formatter.rb
@@ -1,11 +1,11 @@
 # frozen_string_literal: true
 
-require "trace_viz/formatters/base_formatter"
+require "trace_viz/formatters/trace_data_formatter"
 
 module TraceViz
   module Exporters
     module Formatters
-      class BaseFormatter < TraceViz::Formatters::BaseFormatter
+      class BaseFormatter < TraceViz::Formatters::TraceDataFormatter
       end
     end
   end
