+++ lib/trace_viz/loggers/trace_formatters/base_formatter.rb
@@ -1,50 +1,17 @@
 # frozen_string_literal: true
 
-require "trace_viz/formatters/base_formatter"
-require "trace_viz/utils/colorize"
+require "trace_viz/formatters/trace_data_formatter"
+require_relative "helpers/color_helper"
+require_relative "helpers/depth_helper"
+require_relative "helpers/method_name_helper"
 
 module TraceViz
   module Loggers
     module TraceFormatters
-      class BaseFormatter < TraceViz::Formatters::BaseFormatter
-        private
-
-        def colorize(text, *styles)
-          Utils::Colorize.colorize(text, *styles)
-        end
-
-        def format_depth
-          return unless config.general[:show_depth]
-
-          prefix = colorize("depth", :dip, :italic, :blue)
-          open_block = colorize("[", :dip, :light_blue)
-          number = colorize(trace_data.depth, :light_red)
-          close_block = colorize("]", :dip, :light_blue)
-
-          [
-            prefix,
-            open_block,
-            number,
-            close_block,
-          ].join
-        end
-
-        def format_method_name
-          return unless config.general[:show_method_name]
-
-          "#{format_class_name}#{format_action_name}"
-        end
-
-        def format_class_name
-          klass = colorize(trace_data.klass, :light_green)
-          method_sign = colorize("#", :blue)
-
-          [klass, method_sign].join
-        end
-
-        def format_action_name
-          colorize(trace_data.action, :bold, :light_cyan)
-        end
+      class BaseFormatter < TraceViz::Formatters::TraceDataFormatter
+        include Helpers::ColorHelper
+        include Helpers::DepthHelper
+        include Helpers::MethodNameHelper
       end
     end
   end
