+++ lib/trace_viz/loggers/trace_formatters/method_return_formatter.rb
@@ -1,40 +1,23 @@
 # frozen_string_literal: true
 
-require "trace_viz/utils/format_utils"
 require_relative "base_formatter"
+require_relative "helpers/result_helper"
 
 module TraceViz
   module Loggers
     module TraceFormatters
       class MethodReturnFormatter < BaseFormatter
+        include Helpers::ResultHelper
+
         def format
           [
             indent_representation,
-            format_depth,
-            format_method_name,
-            formatted_result,
+            format_depth(trace_data, config),
+            format_method_name(trace_data, config),
+            format_result(trace_data, config),
             colorize(source_location_representation, :dark_gray),
             colorize(execution_time_representation, :red),
           ].compact.join(" ")
-        end
-
-        private
-
-        def formatted_result
-          return unless config.result[:show]
-
-          truncated_result = Utils::FormatUtils.truncate_value(
-            trace_data.result.inspect,
-            config.result[:truncate_length],
-          )
-
-          prefix = colorize("#=>", :dip, :italic, :light_blue)
-          result = colorize(truncated_result, :dip, :gray)
-
-          [
-            prefix,
-            result,
-          ].join(" ")
         end
       end
     end
